Doing some BCD decoding for time and date
Going to change the strategy a little on this, by decoding the stream of bits into a stream of digits, which can constructed at the end of the minute.

Noticed the time is about one second behind, so the next second must start on the carrier drop after >-500mS of carrier on. The specs don't make it clear when the minute actually starts.